STYLE: Improve transaction popup styling
This commit improves the styling of the transaction popup window by adding `overflow: auto;` to its CSS class. This ensures that the content of the popup, which may exceed the window height, will have a vertical scrollbar, improving user experience and accessibility.  There were no other changes in this commit.